+++ apps/web/components/widgets/now-playing-widget.tsx
