+++ apps/web/components/widgets/web-vitals-reporter.tsx
